+++ CHANGELOG.rst
@@ -1,9 +1,9 @@
 httsleep Changelog
 ==================
 
-Next release
-------------
-* Replace *args and **kwargs in httsleep wrapper with explicit args/kwargs, so
+Version 0.1.4
+-------------
+* Replace ``*args`` and ``**kwargs`` in httsleep wrapper with explicit args/kwargs, so
   they can be displayed in IDEs
 * Documentation improvements